+++ src/features/search-post/index.ts
@@ -1,0 +1,1 @@
+export { useSearchPosts } from "./api/useSearchPosts";
